Refactor type names JC-637
