+++ apps/angular/src/core/services/anime.service.ts
@@ -1,28 +1,18 @@
 import { inject, Injectable } from '@angular/core';
-
 import { HttpClient } from '@angular/common/http';
-
 import { map, Observable } from 'rxjs';
-
-import { AnimeMapper } from '@js-camp/core/mappers/anime.mapper';
-import { TAnime } from '@js-camp/core/models/anime';
-import { PaginationMapper } from '@js-camp/core/mappers/pagination.mapper';
-import { PaginationDto } from '@js-camp/core/dtos/pagination.dto';
-import { Pagination } from '@js-camp/core/models/pagination';
-import { TAnimeDto } from '@js-camp/core/dtos/anime.dto';
 import { AppUrlsConfig } from '@js-camp/angular/app/shared/app-url';
+import { AnimeDto } from '@js-camp/core/dtos/anime.dto';
+import { PaginationDto } from '@js-camp/core/dtos/pagination.dto';
+import { AnimeMapper } from '@js-camp/core/mappers/anime.mapper';
+import { PaginationMapper } from '@js-camp/core/mappers/pagination.mapper';
+import { Anime } from '@js-camp/core/models/anime';
+import { Pagination } from '@js-camp/core/models/pagination';
 
-/** Anime pagination Dto. */
-type AnimeResponseDto = PaginationDto<TAnimeDto>;
-
-/** Anime pagination. */
-export type AnimeResponse = Pagination<TAnime>;
-
-/** Anime services. */
+/** Anime service. */
 @Injectable({
 	providedIn: 'root',
 })
-
 export class AnimeService {
 	private readonly appUrlsConfig = inject(AppUrlsConfig);
 
@@ -33,8 +23,8 @@ export class AnimeService {
 	private readonly paginationMapper = inject(PaginationMapper);
 
 	/** Return an observable containing the list of anime. */
-	public getAllAnime(): Observable<AnimeResponse> {
-		return this.httpClient.get<AnimeResponseDto>(this.appUrlsConfig.anime.list).pipe(
+	public getAllAnime(): Observable<Pagination<Anime>> {
+		return this.httpClient.get<PaginationDto<AnimeDto>>(this.appUrlsConfig.anime.list).pipe(
 			map(responseDto => this.paginationMapper.mapPaginationFromDto(responseDto, this.animeMapper)),
 		);
 	}
