+++ libs/core/dtos/datetime-range.dto.ts
@@ -1,0 +1,9 @@
+/** Date Time Range Type. */
+export type DateTimeRangeDto = {
+
+	/** Start date. */
+	start: string | null;
+
+	/** End date. */
+	end: string | null;
+};
