+++ libs/core/mappers/pagination.mapper.ts
@@ -4,7 +4,7 @@ import { Pagination } from '../models/pagination';
 
 import { PaginationDto } from '../dtos/pagination.dto';
 
-import { TMapperFromDto, TMapperFunction } from '../types/mapper';
+import { TMapperFromDto, TMapperFunction } from '../models/mapper';
 
 /** Pagination Mapper. */
 @Injectable({
