+++ libs/core/models/datetime-range.ts
@@ -1,0 +1,9 @@
+/** Date Time Range Type. */
+export type DateTimeRange = {
+
+	/** Start date. */
+	startDate: Date | null;
+
+	/** End date. */
+	endDate: Date | null;
+};
