--- libs/core/types/datetime-range.ts
@@ -1,19 +1,0 @@
-/** Date Time Range Type. */
-export type TDateTimeRange = {
-
-	/** Start date. */
-	startDate: Date | null;
-
-	/** End date. */
-	endDate: Date | null;
-};
-
-/** Date Time Range Type. */
-export type TDateTimeRangeDto = {
-
-	/** Start date. */
-	start: string | null;
-
-	/** End date. */
-	end: string | null;
-};
